refactor: entity 디렉터리 수정
[#53]
