+++ packages/backend/src/configs/typeorm.config.ts
@@ -3,6 +3,6 @@ import { TypeOrmModuleOptions } from '@nestjs/typeorm';
 export const typeOrmConfig: TypeOrmModuleOptions = {
   type: 'sqlite',
   database: 'db.sqlite',
-  entities: [__dirname + '/../**/*.entity.{js,ts}'],
+  entities: [__dirname + '/../**/entity/*.entity.{js,ts}'],
   synchronize: true,
 };
