+++ packages/backend/src/problems/problems.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ProblemsController } from './problems.controller';
 import { ProblemsService } from './problems.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Problem } from './problem.entity';
+import { Problem } from './entity/problem.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Problem])],
