+++ packages/backend/src/problems/problems.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Problem } from './problem.entity';
+import { Problem } from './entity/problem.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
